minimal friendship works ... so far, a user can query another user, send friend request, can cancel that request, accept and decline that request as well as unfriend a player.

+++ source/transcendence/friendship/views.py
@@ -15,6 +15,7 @@ from rest_framework.exceptions import AuthenticationFailed
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 import urllib.parse
+import json
 
 
 # template_name for viewing player profile
@@ -36,9 +37,6 @@ class PlayerProfileView(APIView, BaseView):
 			response.delete_cookie('refresh_token')
 			return response
 		return super().handle_exception(exception)
-	# def get(self, request, *args, **kwargs):  # Use standard args and kwargs
-	# 	print('PlayerProfileView - from friendship app', flush=True)
-	# 	return super().get(request, *args, **kwargs)  # Pass both args and kwargs
 
 	def get_player(self, username):
 		return Player.objects.filter(username=username).first()
@@ -81,61 +79,97 @@ class FriendRequestView(APIView):
 	authentication_classes = [JWTCookieAuthentication]
 	permission_classes = [IsAuthenticated]
 
+	def get_player(self, username):
+		return Player.objects.filter(username=username).first()
+	
 	# send friend request - PENDING
 	def post(self, request, *args, **kwargs):
 		print('FriendRequestView - post - to add friend request', flush=True)
 		data = request.data
 		data['sender'] = request.user.id  # Using sender's primary key
 
-		receiver = get_object_or_404(Player, username=kwargs.get('username'))
+		receiver = self.get_player(kwargs.get('username'))
+		if not receiver:
+			print('Friend request reciever Player not found', flush=True)
+			return {'error_msg':'Player not found - cant send friend request'}
+		
 		data['receiver'] = receiver.id
 
 		serializer = FriendRequestSerializer(data=data)
 		if serializer.is_valid():
 			serializer.save()
+			print(f'{request.user.username} SENT FRIEND REQUEST to {receiver.username}!', flush=True)
 			return Response(serializer.data, status=status.HTTP_201_CREATED) # 
-
 		# Return detailed error information
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 	# Cancel friend request - CANCELLED
 	def patch(self, request, *args, **kwargs):
 		print('FriendRequestView - patch - to cancel friend request', flush=True)
-		receiver = get_object_or_404(Player, username=kwargs.get('username'))
+		receiver = self.get_player(kwargs.get('username'))
+		if not receiver:
+			print('Friend cancel request  Player not found', flush=True)
+			return {'error_msg':'Player not found - cant cancel friend request'}
 		friend_req = FriendRequest.objects.filter(sender=request.user.id, receiver=receiver.id).first()
 		if not friend_req:
-			return Response('No friend request found', status=status.HTTP_404_NOT_FOUND)
+			return Response({'error_msg': 'No friend request found'}, status=status.HTTP_404_NOT_FOUND)
+		friend_req.cancel()
 		friend_req.delete()
-		return Response('Friend request CANCELLED', status=status.HTTP_200_OK)
+		print(f'{request.user.username} CANCELLED {receiver.username}\'s request!', flush=True)
+		return Response({'success': 'Friend request CANCELLED'}, status=status.HTTP_200_OK)
+
+	# remove a Friend - un-friend
+	def delete(self, request, *args, **kwargs):
+		print('FriendRequestView - delete - to remove friend', flush=True)
+		receiver = self.get_player(kwargs.get('username'))
+		if not receiver:
+			print('Player doesnt exist - so cant unfriend him', flush=True)
+			return {'error_msg':'Player not found - cant unfriend'}
+		current_players_list = FriendList.objects.get(player=request.user)
+		try:
+			current_players_list.remove_friend(receiver)
+			print(f'{request.user.username} UN_FRIENDED {receiver.username}!', flush=True)
+		except Exception as e:
+			return Response({'error_msg': str(e)}, status=status.HTTP_400_BAD_REQUEST)
+		return Response({'success': 'Friend removed'}, status=status.HTTP_200_OK)
 
 # RESPONDING TO FRIEND REQUEST - FROM RECEIVER PERSPECTIVE
 	""" 
 		we can use PATCH method to accept or decline the friend request  based on what action the user wants to take.
 	"""
 class FriendRequestResponseView(APIView):
+	authentication_classes = [JWTCookieAuthentication]
+	permission_classes = [IsAuthenticated]
+
+	def get_player(self, username):
+		return Player.objects.filter(username=username).first()
+	
 	def patch(self, request, *args, **kwargs):
 		data = request.data
 		data['sender'] = request.user.id
-		receiver = get_object_or_404(Player, username=kwargs.get('username'))
+		receiver = self.get_player(kwargs.get('username'))
+		if not receiver:
+			print('Player doesnt exist - cant respond to friend request', flush=True)
+			return {'error_msg':'Player not found - cant respond to friend request'}
 		data['receiver'] = receiver.id
 		friend_req = FriendRequest.objects.filter(sender=receiver, receiver=request.user).first()
 		if not friend_req:
-			return Response('No friend request found', status=status.HTTP_404_NOT_FOUND)
+			return Response({'error_msg': 'No friend request found'}, status=status.HTTP_404_NOT_FOUND)
 		if data.get('action') == 'decline':
 			friend_req.decline()
-			return Response('Friend request declined', status=status.HTTP_200_OK)
+			friend_req.delete()
+			print(f'{request.user.username} DECLINED {receiver.username}\'s request!', flush=True)
 		elif data.get('action') == 'accept':
+			print(f'{request.user.username} ACCEPTED {receiver.username}\'s request!', flush=True)
+			# we will either update the satatus or delete it coz its fulfilled
 			friend_req.accept()
+			friend_req.delete()
 			# after accepting
 			""" 
 			 1. add each other to their friend list
 			 2. remove the friend request from the friend table
 			"""
-			other_players_list = FriendList.objects.get(player=receiver)
-			current_players_list = FriendList.objects.get(player=request.user)
-			other_players_list.friends.add(request.user)
-			current_players_list.friends.add(receiver)
-			friend_req.delete()
-			return Response('Friend request accepted', status=status.HTTP_200_OK)
-
+		friend_list = FriendList.objects.get(player=request.user)
+		print('Friend list: ', friend_list, flush=True)
+		return Response({'success': 'Friend request fulfilled'}, status=status.HTTP_200_OK)
